observability/genaisdk/requirements.txt
# Note: the versions listed here are merely the ones that this was
# tested with when the sample code was written. You will probably
# want to bump the version numbers to something more recent when
# you write your own code based on this sample code.
google-auth==2.38.0
google-genai==1.7.0
google-cloud-logging==3.11.4
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation-google-genai==0.1b0
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-exporter-gcp-logging==1.9.0a0
opentelemetry-exporter-gcp-monitoring==1.9.0a0
python-dotenv==1.1.0
requests==2.32.3
